% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locf}
\alias{locf}
\title{Last Observation Carried Forward}
\usage{
locf(x, rule = 0.75)
}
\arguments{
\item{x}{Dataframe}

\item{rule}{What proportion of values need to be present before imputing? Defaults to 0.75}
}
\description{
This function allows you to impute missing values with the previous observation.
}
\examples{
locf()
}
\keyword{imputation}
