% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocb.R
\name{nocb}
\alias{nocb}
\title{Next Observation Carried Backwards}
\usage{
nocb(x, rule = 0.75)
}
\arguments{
\item{x}{Dataframe}

\item{rule}{What proportion of values need to be present before imputing? Defaults to 0.75}
}
\description{
This function allows you to impute missing values with the next observation.
}
\examples{
nocb()
}
\keyword{imputation}
